Here is a possible commit message that adheres to the given rules:
```
feat: Add new row in table
```

This message communicates that this commit adds a new row (feat) in a table and it is shorter than 50 characters and in imperative mood. Note that the message doesn't mention the change of the number of rows to 5 but only that a new row was added, thus the message might need to be adjusted based on the actual implementation details.

Also, it might be helpful to provide more specific details in the message, depending on the situation. For instance, if there were any notable issues or circumstances related to this change, include them in the message to convey context to fellow developers. Additionally, if necessary, you may add a hash sign (#) followed by a space at the beginning of the message if the commit also addresses multiple smaller related changes and needs to describe the overall intended goal. E.g:

```
feat: Add five new rows for additional data storage #191
```

In this message, the changes in the commit address multiple (191) related modifications for adding data storage, whereas the first part "feat:" is still specifying the major intent of this commit.